Update README.md
Let’s rewind a bit. Picture this: you’re in Australia, soaking up the sun and studying to be an architect. You’re dreaming of designing skyscrapers, but somehow you end up designing ships and submarines. Because, why not? It’s like architecture, but wetter.
Fast forward a few years and you decide to trade the sunny shores of Oz for the charming drizzle of the UK. Because who wouldn’t swap Bondi Beach for a good cup of tea? In between dodging sheep and navigating roundabouts, you discover a few things: a love for foraging (because nature’s grocery store is free), painting (because blank canvases don’t talk back), and coding (because apparently, designing submarines wasn’t complicated enough).
Somewhere along the line, you realize that while submarines are cool, the world of tech is calling your name. So, you dive headfirst into full stack development, because what could go wrong when you mix a love for nature with lines of code? Spoiler alert: a lot, but it’s all part of the fun.
These days, when you’re not elbow-deep in code, you’re probably out in the woods, pretending to be a contestant on a very low-budget survival show, or painting landscapes that may or may not have kangaroos hidden in them (force of habit).
So, what exactly do I bring to the table? A quirky mix of leadership, software engineering, web development, APIs, testing, full stack wizardry, product management, and an uncanny ability to spot the difference between a mushroom that’s good in a risotto and one that will make you see unicorns.
In short, I design, I code, I paint, I forage, and I occasionally look at a ship and think, “I used to design those.” All in a day’s work.